Move toPaymentInfo from helpers to core/utils
toPaymentInfo is a PaymentInfo converter, same category as fromAbiPaymentInfo,
parsePaymentInfo, and toAbiPaymentInfo already in core/utils. Helpers should
only contain merchant-server utilities (refundable).
