+++ packages/helpers/src/index.ts
@@ -6,5 +6,3 @@
 
 export { refundable } from "./refundable.js";
 export type { EscrowExtra, PaymentOption, RefundableOptions } from "./types.js";
-export { toPaymentInfo } from "./bridge.js";
-export type { EscrowPayloadLike } from "./bridge.js";
